Create ConfigTool.cmd
Windows batch file which helps swap between multiple saved SRS client configurations, such as A10, F16, F18, etc.  If the current config does not exist, it will ask to save the config to the file named with the airframe extension.

It is intended that you create a shortcut with running the configtool.cmd and the desired airframe file extension as a command line parameter.